move internal directives to private space, cleanup unnecessary docs [ci:force]
