+++ lib/process-services-cloud/src/lib/task/public-api.ts
@@ -22,6 +22,5 @@ export * from './models/task.model';
 export * from './models/start-task-cloud-request.model';
 export * from './task-header/public-api';
 export * from './task-form/public-api';
-export * from './directives/public-api';
 export * from './services/task-cloud.service';
 export * from './task-cloud.module';
