+++ lib/process-services-cloud/src/lib/task/task-form/components/user-task-cloud-buttons/claim-task/claim-task-cloud.directive.ts
@@ -16,14 +16,15 @@
  */
 
 import { Directive, Input, HostListener, Output, EventEmitter, OnInit, ElementRef, Renderer2 } from '@angular/core';
-import { TaskCloudService } from '../services/task-cloud.service';
+import { IdentityUserService } from '../../../../../people/services/identity-user.service';
+import { TaskCloudService } from '../../../../services/task-cloud.service';
 
 @Directive({
     // eslint-disable-next-line @angular-eslint/directive-selector
-    selector: '[adf-cloud-unclaim-task]',
+    selector: '[adf-cloud-claim-task]',
     standalone: true
 })
-export class UnClaimTaskCloudDirective implements OnInit {
+export class ClaimTaskCloudDirective implements OnInit {
     /** (Required) The id of the task. */
     @Input()
     taskId: string;
@@ -42,7 +43,12 @@ export class UnClaimTaskCloudDirective implements OnInit {
 
     invalidParams: string[] = [];
 
-    constructor(private readonly el: ElementRef, private readonly renderer: Renderer2, private taskListService: TaskCloudService) {}
+    constructor(
+        private readonly el: ElementRef,
+        private readonly renderer: Renderer2,
+        private taskListService: TaskCloudService,
+        private identityUserService: IdentityUserService
+    ) {}
 
     ngOnInit() {
         this.validateInputs();
@@ -71,9 +77,20 @@ export class UnClaimTaskCloudDirective implements OnInit {
     @HostListener('click')
     async onClick() {
         try {
+            await this.claimTask();
+        } catch (error) {
+            this.error.emit(error);
+        }
+    }
+
+    private async claimTask() {
+        const currentUser: string = this.identityUserService.getCurrentUserInfo().username;
+        try {
             this.renderer.setAttribute(this.el.nativeElement, 'disabled', 'true');
-            await this.taskListService.unclaimTask(this.appName, this.taskId).toPromise();
-            this.success.emit(this.taskId);
+            const result = await this.taskListService.claimTask(this.appName, this.taskId, currentUser).toPromise();
+            if (result) {
+                this.success.emit(result);
+            }
         } catch (error) {
             this.renderer.removeAttribute(this.el.nativeElement, 'disabled');
             this.error.emit(error);
