+++ lib/process-services-cloud/src/lib/task/task-form/components/user-task-cloud-buttons/unclaim-task/unclaim-task-cloud.directive.spec.ts
@@ -17,24 +17,26 @@
 
 import { Component, ContentChildren, ViewChild } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { TaskCloudService } from '../services/task-cloud.service';
+import { TaskCloudService } from '../../../../services/task-cloud.service';
 import { of, throwError } from 'rxjs';
-import { ClaimTaskCloudDirective } from './claim-task-cloud.directive';
-import { taskClaimCloudMock } from '../task-header/mocks/fake-claim-task.mock';
-import { ProcessServiceCloudTestingModule } from '../../testing/process-service-cloud.testing.module';
+import { UnClaimTaskCloudDirective } from './unclaim-task-cloud.directive';
+import { taskClaimCloudMock } from '../../../../task-header/mocks/fake-claim-task.mock';
+import { ProcessServiceCloudTestingModule } from '../../../../../testing/process-service-cloud.testing.module';
 import { By } from '@angular/platform-browser';
 
-describe('ClaimTaskCloudDirective', () => {
+describe('UnClaimTaskCloudDirective', () => {
     @Component({
-        selector: 'adf-cloud-claim-test-component',
-        template: '<button adf-cloud-claim-task [taskId]="taskMock" [appName]="appNameMock" (error)="onError($event)"></button>'
+        selector: 'adf-cloud-test-component',
+        standalone: true,
+        imports: [UnClaimTaskCloudDirective],
+        template: '<button adf-cloud-unclaim-task [taskId]="taskIdMock" [appName]="appName" (error)="onError($event)"></button>'
     })
     class TestComponent {
-        taskMock = 'test1234';
-        appNameMock = 'simple-app';
+        appName = 'simple-app';
+        taskIdMock = '1234';
 
-        @ViewChild(ClaimTaskCloudDirective, { static: false })
-        claimTaskDirective: ClaimTaskCloudDirective;
+        @ContentChildren(UnClaimTaskCloudDirective)
+        unclaimTaskDirective: UnClaimTaskCloudDirective;
 
         onError(error: Error) {
             return error;
@@ -46,44 +48,43 @@ describe('ClaimTaskCloudDirective', () => {
 
     beforeEach(() => {
         TestBed.configureTestingModule({
-            imports: [ProcessServiceCloudTestingModule],
-            declarations: [TestComponent]
+            imports: [ProcessServiceCloudTestingModule, TestComponent]
         });
         taskCloudService = TestBed.inject(TaskCloudService);
         fixture = TestBed.createComponent(TestComponent);
         fixture.detectChanges();
     });
 
-    it('should directive claim task', () => {
-        spyOn(taskCloudService, 'claimTask').and.returnValue(of(taskClaimCloudMock));
+    it('should directive unclaim task', () => {
+        spyOn(taskCloudService, 'unclaimTask').and.returnValue(of(taskClaimCloudMock));
         const button = fixture.nativeElement.querySelector('button');
         button.click();
-        expect(taskCloudService.claimTask).toHaveBeenCalled();
+        expect(taskCloudService.unclaimTask).toHaveBeenCalled();
     });
 
     it('should emit error on api fail', async () => {
         const error = { message: 'task key not found' };
-        spyOn(taskCloudService, 'claimTask').and.returnValue(throwError(error));
+        spyOn(taskCloudService, 'unclaimTask').and.returnValue(throwError(error));
         spyOn(fixture.componentInstance, 'onError').and.callThrough();
         const button = fixture.nativeElement.querySelector('button');
         button.click();
         await fixture.whenStable();
-        expect(taskCloudService.claimTask).toHaveBeenCalled();
+        expect(taskCloudService.unclaimTask).toHaveBeenCalled();
         expect(fixture.componentInstance.onError).toHaveBeenCalledWith(error);
     });
 
     it('should DISABLE the button on task completion', () => {
-        spyOn(taskCloudService, 'claimTask').and.returnValue(of(taskClaimCloudMock));
+        spyOn(taskCloudService, 'unclaimTask').and.returnValue(of(taskClaimCloudMock));
         const button = fixture.debugElement.query(By.css('button')).nativeElement;
 
         button.click();
 
-        expect(taskCloudService.claimTask).toHaveBeenCalled();
+        expect(taskCloudService.unclaimTask).toHaveBeenCalled();
         expect(button.disabled).toBe(true);
     });
 
     it('should ENABLE the button on api failure', () => {
-        spyOn(taskCloudService, 'claimTask').and.throwError('process key not found');
+        spyOn(taskCloudService, 'unclaimTask').and.throwError('process key not found');
         const button = fixture.debugElement.query(By.css('button')).nativeElement;
 
         button.click();
@@ -92,61 +93,69 @@ describe('ClaimTaskCloudDirective', () => {
     });
 });
 
-describe('Claim Task Directive validation errors', () => {
+describe('UnClaim Task Directive validation errors', () => {
     @Component({
         selector: 'adf-cloud-claim-no-fields-validation-component',
-        template: '<button adf-cloud-claim-task></button>'
+        standalone: true,
+        imports: [UnClaimTaskCloudDirective],
+        template: '<button adf-cloud-unclaim-task></button>'
     })
     class ClaimTestMissingInputDirectiveComponent {
         appName = 'simple-app';
         appNameUndefined = undefined;
         appNameNull = null;
 
-        @ContentChildren(ClaimTaskCloudDirective)
-        claimTaskValidationDirective: ClaimTaskCloudDirective;
+        @ContentChildren(UnClaimTaskCloudDirective)
+        claimTaskValidationDirective: UnClaimTaskCloudDirective;
     }
 
     @Component({
         selector: 'adf-cloud-claim-no-taskid-validation-component',
-        template: '<button adf-cloud-claim-task [appName]="appName"></button>'
+        standalone: true,
+        imports: [UnClaimTaskCloudDirective],
+        template: '<button adf-cloud-unclaim-task [appName]="appName"></button>'
     })
     class ClaimTestMissingTaskIdDirectiveComponent {
         appName = 'simple-app';
 
-        @ContentChildren(ClaimTaskCloudDirective)
-        claimTaskValidationDirective: ClaimTaskCloudDirective;
+        @ContentChildren(UnClaimTaskCloudDirective)
+        claimTaskValidationDirective: UnClaimTaskCloudDirective;
     }
 
     @Component({
         selector: 'adf-cloud-claim-undefined-appname-component',
-        template: '<button adf-cloud-claim-task [taskId]="taskMock" [appName]="appNameUndefined"></button>'
+        standalone: true,
+        imports: [UnClaimTaskCloudDirective],
+        template: '<button adf-cloud-unclaim-task [taskId]="taskMock" [appName]="appNameUndefined"></button>'
     })
     class ClaimTestInvalidAppNameUndefinedDirectiveComponent {
         appNameUndefined = undefined;
         taskMock = 'test1234';
 
-        @ContentChildren(ClaimTaskCloudDirective)
-        claimTaskValidationDirective: ClaimTaskCloudDirective;
+        @ContentChildren(UnClaimTaskCloudDirective)
+        claimTaskValidationDirective: UnClaimTaskCloudDirective;
     }
 
     @Component({
         selector: 'adf-cloud-claim-null-appname-component',
-        template: '<button adf-cloud-claim-task [taskId]="taskMock" [appName]="appNameNull"></button>'
+        standalone: true,
+        imports: [UnClaimTaskCloudDirective],
+        template: '<button adf-cloud-unclaim-task [taskId]="taskMock" [appName]="appNameNull"></button>'
     })
     class ClaimTestInvalidAppNameNullDirectiveComponent {
         appNameNull = null;
         taskMock = 'test1234';
 
-        @ViewChild(ClaimTaskCloudDirective, { static: false })
-        claimTaskValidationDirective: ClaimTaskCloudDirective;
+        @ViewChild(UnClaimTaskCloudDirective, { static: false })
+        claimTaskValidationDirective: UnClaimTaskCloudDirective;
     }
 
     let fixture: ComponentFixture<any>;
 
     beforeEach(() => {
         TestBed.configureTestingModule({
-            imports: [ProcessServiceCloudTestingModule],
-            declarations: [
+            imports: [
+                ProcessServiceCloudTestingModule,
                 ClaimTestMissingTaskIdDirectiveComponent,
                 ClaimTestInvalidAppNameUndefinedDirectiveComponent,
                 ClaimTestInvalidAppNameNullDirectiveComponent,
